Refactor to use NoGhcTc.
In a few places in the AST, we need to store a type in a non-type
AST, like in type annotations, type applications, and pattern
annotations. In all of these places, we need a, e.g.,
HsExpr GhcTc to store an HsType GhcRn -- note that the type never
makes it past GhcRn. Before this patch, this was done via the extension
fields in the AST. This is unsatisfying, though, because the e.g.
type in an annotated expression is surely not an extension, but the
raison d'être for the AST node in the first place.

This patch introduces a new type family NoGhcTc such that
NoGhcTc GhcTc = GhcRn and behaves like the type-level identity
otherwise. This new type family is used in a few places that
benefit from it. There should be no outward change in behavior,
and hence, no tests.