More explicit foralls (GHC Proposal 0007)
Allow the user to explicitly bind type/kind variables in the following
places:
- type/newtype/data instances (including associated instances)
- closed type family equations
- RULES pragmas
when the ExplicitForAll extension is enabled. Extend -Wunused-foralls
and -Wunused-type-patterns to handle these cases, include appropriate
tests, and update the users' guide. Changes to the Template Haskell
AST are documented in libraries/template-haskell/changelog.md.

Also introduce the type family NoGhcTc to HsExtension.hs, useful when
we need to store a type (which never makes it past GhcRn) in a
non-type AST (which does make it to GhcTc). We define NoGhcTc GhcTc =
GhcRn and NoGhcTc to be the type-level identity otherwise. (Refactors
HsAppType, ExprWithTySig, and SigPat.)

Other Changes:
- Convert HsRule to a record
- Add repMaybe to DsMeta

Follows the specification of GHC Proposal 0007, also fixes Trac #2600.
Advised by Richard Eisenberg.